Rename the playbooks to include proper tox targets.
Currently it's confusing to use functional playbook
if it in fact calls dsvm-functional tox target.

This change renames the playbook that is required
by tests using dsvm-functional tox target and
keeps the functional playbook for functional tox
target.

If required additional playbooks should be created
to match the tox target.

Change-Id: I77d6f520c1f61bfdf9ed5ee8a1db384aefa49ac9