Add common note to README for hosts file
add example for changing the username for the hosts entry
since this is a common example where you may need to log
into your tester using a different user.  Let's add this to
the example to make it easier for people to learn as they go along.

this is common knowledge for ansible users but not so much for people
just learning ansible and octario

Change-Id: I6538d8c45281d47a173898c4cf0bc33bf1c4431c