Update patch set 21
Patch Set 21:

(1 comment)

Reading the logs, I think it happens because the virtualenv does not use system packages and tries to install again what is not needed. But for sure this is not required when the tempest RPM is used (as all dependencies are provided); the system packages should be enough also when redhat-openstack/tempest is used from git.

Patch-set: 21
Label: Code-Review=0