neutron: load bridge firewalling kernel module for functional tests
This is needed because some functional tests (for iptables firewall) set
sysctl knobs that are provided by those kernel modules.

In newer kernels (3.18+), the knobs were moved into a separate kernel
module (br_netfilter).

Technically, we don't need to load bridge module because it will be
loaded by the first brctl invocation which will happen before an
iptables rule for a bridge is set by a test, so loading the module is
merely for completeness.

This patch should solve some of functional test failures when running on
newer platforms (RHEL 7.3+).

Change-Id: Ia0e9b519ccca178e64ac3977ffb72fd9e472e5a4