Update patch set 1
Patch Set 1: Code-Review-2

I thought it's a system config file, just noticed that it's actually the component code, so a big -2 for that.

I would expect to see a patch that remove these sed lines, unless there is a very good reason to run those, that accepted by the component cores.

Patch-set: 1