Update patch set 1
Patch Set 1:

I totally agree that use of sed is malefic but that's only because the original jenkins-run.sh file was using it and I followed the anti-pattern when moving the configuration to octario.

If everyone agrees we should drop sed usage completely but this would require making these changes directly into all patch branches.

I am waiting for Elise to comment on that.

Patch-set: 1
Label: Code-Review=0