Generic way to fetch logs from common directories
This change finds the *.log files in the common test
directories such as .tox and ../logs and merge the list
of found log files with the ones specified by the component
configuration files.

The merged list (unified - no duplicated files are listed) is then
used to fetch the log files.

This will allow to remove most of the archive sections.

Change-Id: I9c0ba1554abc35d2a4d01206411dcac4bbc7d7fe