Removal of archive section from components.
Generic mechanism to collect common log files allows
to remove archive section from most of the config files.

This removes that section together with required changes
for the file names required for the generic mechanism to work,
such as extension to be a .log rather then .txt.

Cosmetic comments changes to reflect above.