Upgraded human_readable output callback
"debug" callback was backported from Ansible 2.2 and once we upgrade to 2.2 we can remove it from our codebase. The debug callback is mainly the new version of human_callback which was finally merge to ansible core! Do not be confused by its name, it has almost nothing to do with debug: it just prints nicely stdout and stderr.

Change-Id: I108ad227d7f5add311ae71faa54cfbe72ff43888