Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (1 comment)
> 
> There are several pep8 failures in debug.py.
> 
> Can you please provide link to poc build to show how it looks and to verify it works?

This is what I was doing right now: building a jenkins job that tests the new output. In fact any execution should validate it. 

Now the tricky part is that debug.py code was copied unmodified from ansible so it does not make any sense to lint it. I guess the solution is to add it as an exception for linting.

Patch-set: 1