Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Any reason why we shouldn't lint it? I have no idea why Ansible folks merged it this way, but if we can fix it, shouldn't we do it?

I think it would be a waste of time to fix it here because this code will be removed as soon we switch to 2.2.

Yes, we can try to fix it in ansible directly.

BTW, what command have you used to do the linting for octario? To my surprise tox is configured to do only ansible-linting, no pep8 or flake8 execution which is a bit sad.

Do you want me to enable flake8 as part of this changeset?

Patch-set: 1
Label: Code-Review=0