Update patch set 2
Patch Set 2:

The task names are there, you didn't see any because my tests were calling the shell module directly, without even having a playbook.

Here is the output from a playbook:

http://pastebin.test.redhat.com/448762

As you can see the only thing changes is the fact that stderr, stdout and msg (when they exist) are removed from JSON code and printed outside in a human friendly way. 

That's perfect for parsing the console log!

Patch-set: 2