puppet-tripleo: use --full-index option when running bundle install
The version of rubygems-bundler provided by downstream CI is:
rubygem-bundler-1.7.8-3.el7.noarch

But it's too old. We hit this bug:
https://github.com/bundler/bundler/issues/3367

We should at least have 1.7.13 but I don't know how to upgrade it.
In the meantime, let's workaround it by adding the option --full-index
option which should fix the situation seen here:
https://paste.fedoraproject.org/paste/2nHjn4MlAfCiJI473HEz615M1UNdIGYhyRLivL9gydE=

Change-Id: I63422d97949d1e79b2db8161f1ba01a8554f9012