Adds nested 'description' key in plugin.spec
Because InfraRed CI runs gates which using this plugin, this patch only
added the new nested description key that that needed for the main
change in InfraRed main project.
The root level 'description' and the 'help' will be removed after the
main change in the InfraRed repo will be merged.

RHOSINFRA-568

Change-Id: If41758677f9e0aab9b741c1dc8a99d23d87f75a2