small tweak to better retrieve correct repo_url
use Git logic to check rhos, patches and then origin as remotes
to fetch repo_url from.  this is because it is common to use
a dist-git + patches remote setup dev/packaging work and
in that case origin points to the dist-git repo normally and
you need to either look at patches or rhos remote.

Change-Id: I11bc1a1c32f80a62b1c9e1d7978c3414161654aa