Update patch set 2
Patch Set 2:

I really don't think there is a need to explain why ignore_errors is evil: because it does HIDE ERRORS in execution. It is a bad practice "aka poor programming" to hide errors. 

It hides any kind of errors, even if the ansible module itself fails.

One side effect from ignore_errors is that Ansible will display them as successes *green* instead of *red*, making much harder to investigate the logs.

The number of ignore_errors in our code is huge and I don't have time or energy to fix it but when I am affected directly by a bug, I fix it.

Here is the job that used this CR:
http://butler.usersys.redhat.com/job/DFG-storage-swift-7-unit-rhos/28/

Patch-set: 2
Label: Code-Review=0