Update patch set 1
Patch Set 1:

I think we have a requirement for a new phase here.\

So the part of octario flow would be:

pre_install_steps: ...
install packages: ...
run:...

this pre_install_steps would be a new role which will require an update of all the playbooks and should work in case pre_install_steps is empty.

I know it's more general octario task, so basically, we can do it, but right now we are handling several issues in CI, so I can say for sure when it will be ready. My estimation is 2-3 days.

Patch-set: 1