Update patch set 1
Patch Set 1:

rhelosp-12.0-brew is needed in order to get access to python2-osc-lib-tests which is not in -unittest repo -- don't ask me why is not there, Jon should be able to provide more info on that.

I do agree with you that we need to test, that's why I marked this as "wip" so we can all test without merging it.

Thanks for the quick feedback on it.

Patch-set: 1
Reviewer: rhosqeauto <1008131@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0