Update patch set 2
Patch Set 2: Code-Review-2

This change is removing the task that tries to auto-detect where the component is without explicitly mentioning it. It ignores completely the assumptions we made when creating the project like not everyone are using octario via CLI.

Let's focus on patches that bring value to our teams.

Patch-set: 2
Label: Code-Review=-2