Update patch set 3
Patch Set 3:

Arie, I am not sure I understand the question but I will try to answer. Currently it was impossible to tell octario where to save the collected files. It was always in the CWD.

This new *optional* parameter, allows user to tell octario where to collect the logs. When not specified it will use CWD, keeping the old behaviour.

In addition to that this assures that this feature works identically when octario is called directly or via infrared.

This was implemented by matching the identical parameter found on infrared/plugins/collect_logs in order to avoid confusions.

Patch-set: 3
Reviewer: Sorin Sbarnea <1008329@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0