Update patch set 2
Patch Set 2: Code-Review-1

Clearly osc-lib-tests is not required by everyone and we already have successful OSP 12 builds for some components so why are we adding a workaround for all components when only some of them require it?

This makes the generic roles look messy and this is exactly why we have the components configuration.

Please add it for components who use it and not for everyone.

Thanks

Patch-set: 2
Reviewer: Arie Bregman <1004788@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1