Update patch set 2
Patch Set 2:

Arie, first, this does only enable the repository, it does *not* install any additional RPM.

I came to this approach because it requires less maintenance in long term. 

Jon can add extra info but from what I talked with him it seems to be a design issue: if we add some new *test* RPMs on a new RHOS version, our jobs would not be able to consume it before the release, as these are found only inside brew until the release.

By activating brew repository before the release we allow to address this issue.

Please read https://projects.engineering.redhat.com/browse/RHOS-23894 which was open for 4 months and is still not fixed. 

Who knows, maybe we can find a better solution, one that would also work for future versions and for other packages. There is a clear trend in moving tests out of the main packages so we will see more of this unless we solve the root cause.

Patch-set: 2
Reviewer: Sorin Sbarnea <1008329@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0