Update patch set 2
Patch Set 2:

"Arie, first, this does only enable the repository, it does *not* install any additional RPM." - thank you for assuming I have no idea what this change is all about, I'm flattered. But, it's k, I reviewed all the changes and it's not that complex.

I came to this approach because it requires less maintenance in long term " - I have no idea what maintenance you are talking about. maybe you can elaborate. Usually, such changes happen because people don't want to change specific files but I'm not going to assume that's the case.

"Jon can add extra info but from what I talked with him it seems to be a design issue: if we add some new *test* RPMs on a new RHOS version, our jobs would not be able to consume it before the release, as these are found only inside brew until the release." - k, I thought this is why we have the copr repos, but I'm probably missing something.


"Who knows, maybe we can find a better solution, one that would also work for future versions and for other packages. There is a clear trend in moving tests out of the main packages so we will see more of this unless we solve the root cause." - I don't know which solution is the best but I know that enabling repository for every component although it's not required by everyone, is a little bit strange.

Patch-set: 2