Update patch set 2
Patch Set 2:

Thanks for looking into this as well, Sorin!

 > Arie, first, this does only enable the repository, it does *not*
 > install any additional RPM.
 
I wonder about this. I noticed on my tester node that after enabling brew for the tripleo/openstack clients, that most packages get installed from brew instead of the release repo because the versions are more recent in brew - so we're not testing exactly what we should. If that's the case maybe it's better to limit the risk to only the components that absolutely require python-osc-lib-tests?

About Copr, it'd be great if we could get python-osc-lib-tests there in the meantime while waiting for the proper fix. I think Jon mentioned python-osc-lib-tests can get out of sync because of the way it's created, so once the automated tagging works we can hopefully avoid a lot of the issues.

Patch-set: 2
Reviewer: Julie Pichon <1011648@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0