Update patch set 1
Patch Set 1:

OK. the job blew up for osp8 because the package that it relied on was removed from pypi. Since we don't control pypi infrastructure, it only makes sense to stop relying on it the same way we do for other jobs.

I think we have a way to include test only packages that are not necessarily part of OSP repos during test run. We had a copr repo for that. I think the right thing to do is ship pep8 / hacking / whatever else is needed in those copr repos, and then land this fix. What are your thoughts?

Patch-set: 1