Update patch set 4
Patch Set 4:

Not sure I see the value in using RPMs for running PEP8.

In some releases we need to use specific versions of pylint, hacking, pep8, pyflakes, ... and usually we find out we don't have those versions so I go and build them but I really don't see the value in doing that. I understand we can find critical issues when using RPMs for unit tests and functional but PEP8?...

Patch-set: 4