name reports based on tester types
This is needed in order to assure that
we can run and collect multiple testers
from a single jenkins job.

Without this jenkins would override test
results and keep only the last collection
due to filename overlappings.

This should work without job changes
because filenames are included in default
collection patterns.

Change-Id: I6d7310ccebdab9010f08c22b3bb8301d797acd2a