Update patch set 2
Patch Set 2:

This change confirmed as working while it was included in https://review.gerrithub.io/398388 CR but we want to merge it before that one in order to make it easier to revert that workaround
later.

This slowsdown the build a little bit but at least it making it work. As soon we have the 
bug fix release on both infrared master and stable, we can revert it.

Patch-set: 2
Label: Code-Review=0