Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> ...and I just found that you are the one that also made this coupling in IR. why? :(
> Why you are doing coupling between projects like adding infrared to octario testing or adding variables that are clearly Jenkins related. This is a very unhealthy approach.

After some digging, I think it may be possible to avoid using BUILD_URL and still be able to provide a good user experience for the CI user, but it will require extra work.

The trick is to find a way to inform CI which log file contains the error, as the playbook execution produces multiple log files which will be collected by Jenkins, but the files themselves do not tell Jenkins "i am the one with the error".

So, just before failing the playbook we create a placeholder file with a name like `~/logs/.failed-testrun.log`, which would tell jenkins that `testrun.log` log contains a failure. The rule is simple, presence of `.failed-` prefixed files indicates failure. Jenkins can loop over collected artifacts and print a nice error message with the URL to the full log file.

Please note that I already considered moving log files to different folder in order to mark them as failed but this does not work because jenkins is likely to
flatten them on collection, loosing the folder information. 

What do you think about this approach? I think it would extendable to other areas, not being specific to octario.

Patch-set: 8