Update patch set 8
Patch Set 8:

Short answer: I like it.

Personally, I like the approach of not making any assumption on user usage of the project. Meaning we first assume it used as a standalone project and not in Jenkins or with Infrared. This way, when tools replaced or disappear (and they do, e.g. Khaleesi) everything is still working :D and there are other benefits like code is more readable and other small stuff but let's always assume it can be cloned and used without people wondering "wait, why there is a BUILD_URL env in code and why it saved in some path that includes 'artifact' word.

Patch-set: 8