Update patch set 8
Patch Set 8:

BUILD_URL is NOT a dependency on Jenkins yet, it's clear it was added for Jenkins usage. We really don't need to pretend that the usage 'BUILD_URL' and 'artifact' has nothing to do with Jenkins.

And while it's NOT a dependency on Jenkins it is a very strange thing to include in the code of a standalone project. Not my invention, simply a way projects being developed today. If you want to couple it with Jenkins, let's have a discussion in one of our meetings on drivers or plugins for CI systems.
Let's just hope people will not use 'BUILD_URL' in IR as an excuse to add more stuff that related to Jenkins but not a real dependency.

Patch-set: 8