Update patch set 16
Patch Set 16:

David, groovy code is saving the *ansible console* output to file, which is very different than the output of the testing commands (single shell task, which is only tail -n100 to console).

With this change octario will start to save full output of test execution to unique files for each tester. This is essential for debugging because we stopped displaying the full output inside octario.

Also this means that we will no longer need the hack we currently have in most component
definitions where we use "tee ../logs/testrun.log" because now this is done by octario itself.

Still, I didn't want to touch 100s files to remove old tee command, it is much better to allow each
component owner to adopt this at their discretion/speed/preference.

Patch-set: 16
Label: Code-Review=0