tox: assure prefix is unique
Previous code was failing to generate unique
prefix for two reasons: use of quotes
which means that expansion was not happening.

Also, use of USER as fallback for case where
BUILD_TAG was missing was redundant as
username was already included at the first part of
the prefix.

New variant is using PPID which is unique.

Change-Id: I0a8d42271a6c43fecd93840e5479b5b542b1cb98
Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>