Uncouple running tests from log collection
Create new role only for loading configuration.

Running tests will the final step in running a specific tester.

Collect logs is now a standalone role with a pre-requirement
of loading the component configuration (so it knows what to
archive per component and not only generally).

To summarize, this patch will allow us to collect logs no matter
what is the result of running tests.

Change-Id: I158387d95f6bd9c90de261daea309cfb453f0540