Functional: Remove configure_host_for_func_testing call
This patch is removing the "configure_host_for_func_testing" call from
the gate_hook.sh script in the OVN functional tests. This is only needed
for configuring the upstream gate.

The patch also removed the "pip install ovs" line from the functional
tests configuration because we should have it installed from the RPM
packages already.

Change-Id: I31f823f39aaca7a4a51a06d47a4e2a1da4a196ce
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>