Fix octavia-ui unit tests
The path the unit tests are using is '.tox/py27...' for loading
xstatic packages (e.g. Angular) but we are using only system packages
(to use RPMs instead of pip packages) so it fails while trying to locate
the packages in tox venv.

This patch replace the tox path with system packages path.

Change-Id: Iddf50259af41f1b7c85dea0023ec3b727d6cbad9