Update patch set 7
Patch Set 7:

Sure!

The privileged system allows to run operations with elevated permissions (capabilities or similar to running as root). In shell it basically would be the parts that need a sudo
For security reasons, when switching to privileged context the environment is tightly cleaned and configured (python path, log outputs, etc).
So for this issue here we ask for privileged command:
https://opendev.org/openstack/neutron/src/branch/master/neutron/agent/linux/ip_lib.py#L928

These parts of functional test run OK from the git checkout, but when switching ("sudo") to privileged we reset most of the configuration/environment. So python looking for privileged neutron module will find it in system (not in git checkout).

Func tests upstream expect a deployment with devstack so would have the same code, this workarounds the install parts to have as close to packaged RPMS we run.

Old privsep spec for further reading: https://specs.openstack.org/openstack/oslo-specs/specs/liberty/privsep.html

Patch-set: 7
Reviewer: Gerrit User 1009495 <1009495@d5d70762-12d0-45a1-890d-524b12d3f735>