Improve testing
This change introduces the improvements to tox configuration file,
tox.ini, making the testing a bit more meaningful right now.
Also the few noticed flaws were corrected.

First, the requirements processing in setup.py was modified
to utilize the pkg_resources module instead of pip. The unnecessary
module imports were removed and PEP-8 compatibility was ensured.

Next, the main tox environment was divided into subenvironments.
In previous design, a failure in early steps prevented the linters
from actually checking the code and reporting the flaws. Now all
checks are launched separately and we are sure they execute.
It makes it easier to develop it further.

Change-Id: If2240263d71380cf10f5ce55e0b1ccd4ef949cf9