Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> Are all copies of the 16.1 files? Maybe some of them could be symbolic links at the beginning.
> 
> Anyway, if those are copies of the current 16.1, that should be fine as a starting point.

Right. But seems that although we tried in the past we tried to manage configs with symbolic links, apparently there was a fallback to the copy and update upproach for the later releases. 
However, Do you clearly prefer symlinks? If that's the case we can change it.

Patch-set: 4