Revert "[rhos13][cinder] Remove rpm_deps for pep8 tests"
It actually breaks the test it was meant to fix

failed: [cinder-102-1618826130-pep8-tester-0] (item=['{{ rpm_deps }}', 'python-cinder', 'python-os-win']) => {
    "ansible_loop_var": "item",
    "attempts": 3,
    "changed": false,
    "item": [
        "{{ rpm_deps }}",
        "python-cinder",
        "python-os-win"
    ]
}

MSG:

It appears that a space separated string of packages was passed in as an
argument. To operate on several packages, pass a comma separated string
of packages or a list of packages.

And it seems that test was working had started working without
any additional fix.

This reverts commit 8c1bbbed5a65cc9af46206283b7235d73c52f795.